Misc: various fixes, partly for LLVM builds

+++ Source/Runtime/Core/Private/Thread/Task/TaskFiberPool.cpp
@@ -23,11 +23,11 @@ class FTaskFiberChunk : Meta::FNonCopyableNorMovable {
 public:
 #if USE_PPE_ASSERT
     // debug programs often use far more stack without the optimizer :
-    STATIC_CONST_INTEGRAL(size_t, StackSize, 4 * ALLOCATION_GRANULARITY); // 256 kb
+    STATIC_CONST_INTEGRAL(u32, StackSize, 4 * ALLOCATION_GRANULARITY); // 256 kb
 #else
-    STATIC_CONST_INTEGRAL(size_t, StackSize, 2 * ALLOCATION_GRANULARITY); // 128 kb
+    STATIC_CONST_INTEGRAL(u32, StackSize, 2 * ALLOCATION_GRANULARITY); // 128 kb
 #endif
-    STATIC_CONST_INTEGRAL(size_t, Capacity, 64); // <=> 64 * 128 kb = 8 mb (16 for debug),
+    STATIC_CONST_INTEGRAL(u32, Capacity, 64); // <=> 64 * 128 kb = 8 mb (16 for debug),
     STATIC_CONST_INTEGRAL(u64, BusyMask, UINT64_MAX); // all bits set <=> all fibers available
 
     using FCallback = FTaskFiberPool::FCallback;
